Fire, flowing lava, falling sand, double chests, blood; 16-bit blocks and a 256-high world
Groundwork for a much bigger game:
- Block ids and texture layers are 16-bit (chunks store Uint16 blocks, the
  vertex format carries a high byte for the layer), and the world is 256
  blocks high with the sea at 63. Chunks are saved in a new run-length
  format; worlds from earlier versions are widened as they load, and the
  multiplayer protocol version goes up so mixed versions don't meet.

Mechanics:
- Flint and steel lights fires on any block. Fire is animated, needs fuel
  or solid ground, burns flammable blocks away (sometimes into more fire),
  jumps to spots beside wood, wool and leaves, sets off TNT, dies down on
  stone, goes out in the rain and can be punched out. Standing in it burns,
  with flames on screen.
- Lava flows: slowly (every 30 ticks) and three blocks across flat ground;
  it hardens against water into obsidian, cobblestone or stone, and sets
  fire to flammable blocks near it (random ticks, as in Minecraft).
- Flowing water and lava surfaces run downhill and falling liquid runs
  down its sides, so you can see which way it moves.
- Sand and gravel fall as smoothly falling blocks that land as blocks.
- Hitting animals and monsters throws red blood (Options can turn it off).
- Two chests side by side make a 54-slot double chest; breaking one half
  leaves a single chest with its things. Shared in multiplayer.
- Using a bed says "Respawn point set".

+++ src/worldgen.js
@@ -136,7 +136,7 @@ export class WorldGen {
   }
 
   generate(cx, cz) {
-    const blocks = new Uint8Array(CHUNK_VOLUME);
+    const blocks = new Uint16Array(CHUNK_VOLUME);
     const climate = new Uint8Array(512);
     const biomes = new Uint8Array(256);
     if (this.type === 'flat') {
